Adding generate DHCP from text file

+++ upload-macs.cgi
@@ -33,17 +33,16 @@ case ${RETPOST} in
 	3)
 		genError 406 "Could not update the mac file" 5
 esac
-
 GETTYPE=$(parseQueryString ${QUERY_STRING} "type")
 [ -z ${GETTYPE} ] && rm ${TMPFILE} && genError 100 "Insufficient parameter " 1
 checkHosttype "${GETTYPE}"
-[ ${?} -ne 0 ] rm ${TMPFILE} && genError 101 "Invalid ACPC Host type" 2
+[ ${?} -ne 0 ] && rm ${TMPFILE} && genError 101 "Invalid ACPC Host type" 2
 FILE="/acpc/adm/etc/${GETTYPE}"
 [ ! -f ${FILE} ] && touch ${FILE}
 isWrite "${FILE}"
 [ ${?} -ne 0 ] && rm ${TMPFILE} && genError 406 "MAC file for ${GETTYPE} has no write permission" 5
 cp ${TMPFILE} ${FILE}
-rm ${TMPFILE}
+#rm ${TMPFILE}
 
 	initResponse
 	startJSON
